+++ src/lib/theme-config.ts
@@ -1,3 +1,3 @@
-export * from '$lib/themes/misiki'
-// export * from '$lib/themes/kesudi'
+// export * from '$lib/themes/misiki'
+export * from '$lib/themes/fashion'
 // export * from '$lib/themes/varni'
