+++ src/lib/themes/fashion/CollectionsHome2.svelte
