+++ src/lib/themes/fashion/DummyProductCard.svelte
