+++ src/lib/themes/fashion/index.ts
@@ -1,17 +1,22 @@
-// Core components of misiki
+// Components of fashion
 
 import CategoriesHome from '$lib/themes/misiki/CategoriesHome.svelte'
 import CategoriesMobile from '$lib/themes/misiki/CategoriesMobile.svelte'
 import CategoriesMobile2 from '$lib/themes/misiki/CategoriesMobile2.svelte'
 import CategoriesSlider from '$lib/themes/misiki/CategoriesSlider.svelte'
-import CollectionsGeneral from '$lib/themes/misiki/CollectionsGeneral.svelte'
+import CollectionsGeneral from '$lib/themes/fashion/CollectionsGeneral.svelte'
+import CollectionsHome from '$lib/themes/fashion/CollectionsHome.svelte'
+import CollectionsHome2 from '$lib/themes/fashion/CollectionsHome2.svelte'
 import Deals from '$lib/themes/misiki/Deals.svelte'
+import DummyProductCard from '$lib/themes/fashion/DummyProductCard.svelte'
 import Hero from '$lib/themes/misiki/Hero.svelte'
 import HeroBanners from '$lib/themes/misiki/HeroBanners.svelte'
 import MegamenuHorizontal from '$lib/themes/misiki/MegamenuHorizontal.svelte'
 import MegamenuVertical from '$lib/themes/misiki/MegamenuVertical.svelte'
 import PickedBanners from '$lib/themes/misiki/PickedBanners.svelte'
-import ProductsListingPage from '$lib/themes/misiki/ProductsListingPage.svelte'
+import ProductCard from '$lib/themes/fashion/ProductCard.svelte'
+import ProductsListingPage from '$lib/themes/fashion/ProductsListingPage.svelte'
+import TrendingProductsHome from '$lib/themes/fashion/TrendingProductsHome.svelte'
 
 export {
     CategoriesHome,
@@ -19,11 +24,16 @@ export {
     CategoriesMobile2,
     CategoriesSlider,
     CollectionsGeneral,
+    CollectionsHome,
+    CollectionsHome2,
     Deals,
+    DummyProductCard,
     Hero,
     HeroBanners,
     MegamenuHorizontal,
     MegamenuVertical,
     PickedBanners,
+    ProductCard,
     ProductsListingPage,
+    TrendingProductsHome,
 }
