--- src/lib/themes/kesudi/index.ts
@@ -1,15 +1,0 @@
-// Components of kesudi
-
-import CollectionsGeneral from '$lib/themes/kesudi/CollectionsGeneral.svelte'
-import CollectionsHome from '$lib/themes/kesudi/CollectionsHome.svelte'
-import CollectionsHome2 from '$lib/themes/kesudi/CollectionsHome2.svelte'
-import DummyProductCard from '$lib/themes/kesudi/DummyProductCard.svelte'
-import TrendingProductsHome from '$lib/themes/kesudi/TrendingProductsHome.svelte'
-
-export {
-    CollectionsGeneral,
-    CollectionsHome,
-    CollectionsHome2,
-    DummyProductCard,
-    TrendingProductsHome,
-}
